open class per file, TemplateCodeGenerator has its own file
